requirements.txt
asgiref==3.10.0
Django==5.0.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
channels==4.0.0
channels-redis==4.2.0
celery==5.4.0
django-celery-beat==2.6.0
daphne==4.0.0